chore(device-authenticity): split files
